requirements.txt
Flask==2.0.3
gevent==23.9.1
httpx==0.27.0
instructor==1.3.7
JPype1==1.5.0
llama_index==0.10.58
openai==1.37.1
pycryptodome==3.20.0
pydantic==2.8.2
PyYAML==6.0.1
redis==5.0.7
Requests==2.32.3
discord==2.3.2
Werkzeug==2.0.3
llama_index.embeddings.ollama==0.1.2
ujson==5.10.0
loguru==0.7.2